// src/enums.ts
export const enum FRONT_END_SCRIPT {
  DEFAULT = 'DEFAULT',
  CARD_LAYOUT = 'CARD_LAYOUT',
  SHOW_MORE_BTN = 'SHOW_MORE_BTN',
  SEARCH_FORM_INPUT = 'SEARCH_FORM_INPUT',
  MINIBAR = 'MINIBAR',
  COLOR_COLUMN = 'COLOR_COLUMN',
}

export const enum LABEL_LEGEND_ID {
  SMALL = 'small',
  MEDIAN = 'median',
  LARGE = 'large',
  AVERAGE = 'average',
  NO_LABEL = 'noLabel',
}